requirements.txt
aiogram~=3.12.0
requests~=2.32.3
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1